pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/reference_builds/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/reference_builds"]

[project]
name = "reference-builds"
dynamic = ["version"]
description = "Building Reference Datasets + Processing Ancillary Data"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Brodie Alexander", email = "Brodie.A.Alexander@rtx.com" },
  { name = "Tadd Bindas", email = "tadd.bindas@ertcorp.com" },
  { name = "Quercus Hamlin", email = "qhamlin@asrcfederal.com" },
  { name = "Dylan Lee", email = "dylan.lee@entarian.com" },
  { name = "Farshid Rahmani", email = "Farshid.Rahmani@rtx.com" },
]
maintainers = [
  { name = "Brodie Alexander", email = "Brodie.A.Alexander@rtx.com" },
  { name = "Quercus Hamlin", email = "qhamlin@asrcfederal.com" },
  { name = "Dylan Lee", email = "dylan.lee@entarian.com" },
]

dependencies = [
  "python-dotenv==1.1.0",
  "geopandas==1.1.1",
  "boto3==1.40.45",
  "pyarrow==20.0.0",
  "pyiceberg[s3fs,glue,sql-sqlite]==0.9.1",
  "pyprojroot==0.3.0",
  "tqdm==4.67.1",
  "folium==0.20.0",
  "matplotlib==3.10.5",
  "mapclassify==2.10.0",
  "numpy==2.2.6",
  "netCDF4==1.7.2",
  "pandas==2.3.0",
  "polars==1.34.0",
  "shapely==2.1.1",
  "xarray==2025.7.1",
  "rioxarray==0.19.0",
  "rasterio==1.4.3",
  "rustworkx==0.17.1",
  "exactextract==0.2.2",
  "rasterstats>=0.20.0",
]

[dependency-groups]
dev = [
  "pre-commit==3.8.0",
  "ruff==0.11.13",
  "mypy==1.15.0",
  "nbstripout==0.8.1",
  "types-PyYaml",
  "types-requests==2.32.4.20250611",
]
examples = [
    "ipykernel==6.29.5",
    "jupyterlab==4.4.3"
]
tests = [
    "pytest==8.4.1",
    "pytest-cov==6.1.1",
    "astropy==7.1.0"
]

[tool.uv]
default-groups = ["dev", "examples", "tests"]

[tool.ruff]
line-length = 110
exclude = [".csv", "LICENSE", ".tf", ".tfvars"]
lint.select = [
  "F",      # Errors detected by Pyflakes
  "E",      # Error detected by Pycodestyle
  "W",      # Warning detected by Pycodestyle
  "I",      # isort
  "D",      # pydocstyle
  "B",      # flake8-bugbear
  "Q",      # flake8-quotes
  "TID",    # flake8-tidy-imports
  "C4",     # flake8-comprehensions
  "BLE",    # flake8-blind-except
  "UP",     # pyupgrade
  "RUF100", # Report unused noqa directives
]
lint.ignore = [
  # line too long -> we accept long comment lines; black gets rid of long code lines
  "E501",
  # Do not assign a lambda expression, use a def -> lambda expression assignments are convenient
  "E731",
  # allow I, O, l as variable names -> I is the identity matrix
  "E741",
  # Missing docstring in public package
  "D104",
  # Missing docstring in public module
  "D100",
  # Missing docstring in __init__
  "D107",
  # Errors from function calls in argument defaults. These are fine when the result is immutable.
  "B008",
  # __magic__ methods are are often self-explanatory, allow missing docstrings
  "D105",
  # first line should end with a period [Bug: doesn't work with single-line docstrings]
  "D400",
  # First line should be in imperative mood; try rephrasing
  "D401",
  ## Disable one in each pair of mutually incompatible rules
  # We don't want a blank line before a class docstring
  "D203",
  # We want docstrings to start immediately after the opening triple quote
  "D213",
  # Bare except okay for passing
  "E722",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"docs/*" = ["I"]
"tests/*" = ["D"]
"*/__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
disallow_any_unimported = false
warn_unused_configs = true
strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:pyogrio",
    "ignore:The 'shapely.geos' module is deprecated:DeprecationWarning",
    "ignore:The behavior of DataFrame concatenation with empty or all-NA entries is deprecated:FutureWarning",
]
